+++ src/numen/lexer.hpp
@@ -43,7 +43,7 @@ public:
 
   template <typename... Ts> std::optional<std::tuple<Ts...>> peakForward(std::size_t i = 0) {
     return [&]<std::size_t... I>(std::index_sequence<I...>) -> std::optional<std::tuple<Ts...>> {
-      std::array<std::optional<Lexer::Token>, sizeof...(Ts)> p{peak(i + I)...};
+      std::array<std::optional<Lexer::Token>, sizeof...(Ts)> p{peak(static_cast<int>(i + I))...};
       bool ok = std::apply([](auto... q) { return (... && (q && std::holds_alternative<Ts>(q->data))); }, p);
       if (!ok) return std::nullopt;
       return std::tuple<Ts...>{std::get<Ts>(p[I]->data)...};
